+++ ooni/geoip.py
@@ -144,15 +144,6 @@ class UbuntuGeoIP(HTTPGeoIPLookupper):
         probe_ip = m.group(1)
         return probe_ip
 
-class DuckDuckGoGeoIP(HTTPGeoIPLookupper):
-    url = "https://api.duckduckgo.com/?q=ip&format=json"
-
-    def parseResponse(self, response_body):
-        j = json.loads(response_body)
-        regexp = "Your IP address is (.*) in "
-        probe_ip = re.search(regexp, j['Answer']).group(1)
-        return probe_ip
-
 INITIAL = 0
 IN_PROGRESS = 1
 
@@ -164,8 +155,7 @@ class ProbeIP(object):
 
     def __init__(self):
         self.geoIPServices = {
-            'ubuntu': UbuntuGeoIP,
-            'duckduckgo': DuckDuckGoGeoIP
+            'ubuntu': UbuntuGeoIP
         }
         self.geodata = {
             'asn': 'AS0',