+++ ooni/settings.py
@@ -22,6 +22,13 @@ basic:
     # Where OONIProbe should be writing its log file
     # logfile: {logfile}
     # loglevel: WARNING
+    # Will rotate logs daily when "daily" or based on rotate_length
+    #  when "length" is set. null to disable rotation.
+    # rotate: daily
+    # rotate_length: 1M
+    # Sets an upper bound on the number of rotated files, only works when
+    #  "length" is the rotation mode
+    # max_rotated_files: null
     # The maximum amount of data to store on disk. Once the quota is reached,
     # we will start deleting older reports.
     # measurement_quota: 1G
@@ -105,6 +112,9 @@ defaults = {
     "basic": {
         "loglevel": "WARNING",
         "logfile": "ooniprobe.log",
+        "rotate": "daily",
+        "rotate_length": "1M",
+        "max_rotated_files": None,
         "measurement_quota": "1G"
     },
     "privacy": {
@@ -156,6 +166,9 @@ if hasattr(sys, 'real_prefix'):
 
 # These are the the embedded settings
 _SETTINGS_INI = os.path.join(OONIPROBE_ROOT, 'settings.ini')
+# This can be overridden by an environment variable for end to end testing
+if 'OONIPROBE_SETTINGS' in os.environ:
+    _SETTINGS_INI = os.environ['OONIPROBE_SETTINGS']
 
 USR_SHARE_PATH = '/usr/share/ooni'
 VAR_LIB_PATH = '/var/lib/ooni'
@@ -345,6 +358,27 @@ class OConfig(object):
             file_path = os.path.join(target_dir, file_name)
             if os.path.isfile(file_path):
                 return file_path
+
+    def log_info(self):
+        log.msg("Paths")
+        log.msg("-----")
+        log.msg("inputs directory: %s" % self.inputs_directory)
+        log.msg("scheduler directory: %s" % self.scheduler_directory)
+        log.msg("resources directory: %s" % self.resources_directory)
+        log.msg("decks available directory: %s" % self.decks_available_directory)
+        log.msg("decks enabled directory: %s" % self.decks_enabled_directory)
+        log.msg("measurements directory: %s" % self.measurements_directory)
+
+        log.msg("running path: %s" % self.running_path)
+        log.msg("config files: %s" % ','.join(self.config_files))
+        log.msg("ooni home: %s" % self.ooni_home)
+        log.msg("var_lib path: %s" % VAR_LIB_PATH)
+        log.msg("etc path: %s" % ETC_PATH)
+        log.msg("usr_share path: %s" % USR_SHARE_PATH)
+
+        log.debug("nettest directory: %s" % self.nettest_directory)
+        log.debug("web ui directory: %s" % self.web_ui_directory)
+
 
     def set_paths(self):
         self.nettest_directory = os.path.join(OONIPROBE_ROOT, 'nettests')